Make forms more consistent
While the main focus of this change is to improve the login screen, part
of the problem was the inconsistent ways that the login and item edit
forms were handled. By using the same approach for both (with the except
of the Remember Me checkbox for the login), the look is much better. The
templates are also easier to understand.