Improve diplay of items with quantities
When someone wants more than one of something, but not an unlimited
amount, the quantity will now be displayed next to the item's name.

Also the claim button will only displayed for items that have an amount
left to claim. Note that claiming an item will cause it to appear to
others as if all of that item has been claimed even though the database
only reflects a claim of one.