Add validator to provide default values
Rather than utilizing a `validate_quantity` method on the Form class
(which apparently isn't called when the field is optional), this
validator can be used with any field to provide a default value.

Why not just use the `default` argument on the field itself? The
quantity should appear to be optional but should default to 1 when no
value is provided. This is a user experience thing. There might be a
better way to do it but for now this is good enough.