Allow users to claim specific quantities
Well I'm not so proud of a lot of the code here, but this is a much
needed feature. I can come back and clean it up later.

The basic idea is that when someone asks for multiple of something (but
not unlimited), a user should be able to claim any number of the item
that is still available. Allowing this to happen requires some changes
to `Item.claim()`, the `frontend.events.claim` view, the JavaScript, and
the template, and the CSS.

And boy do I need to start writing tests for my models.