Use in-memory database for test suite
While it is probably better to use Postgres for the test suite since
that is what is being used for both production and development, for the
time being an in-memory SQLite database will suffice. In order to make
it work, though, the database tables need to be created and destroyed
before each test that uses a request context.

There is probably a better way to do this. I should figure out what that
is.