Add tests for `Item.claim`
Normally I like to try to commit all of a model's tests together, but I
must walk away from the computer now and I'd like to have the tests for
`xmas.events.models.Item.claim()` preserved. This is another important
method to get under test.

When I return to writing tests I need to address the `TODO` about
`item.quantity_claimed`. It doesn't appear to be updated correctly. I'm
not sure yet if it's failing to update in the database or if it's the
instance isn't updating correctly. The update does, after all, go
directly to the database.