Update Item.claim tests to commit before updating
When an item is first entered into the database it has no claims
against it. The original version of the tests for
`xmas.events.models.Item.claim()` treated the instance the same way, but
failed to persist the data failed to the database until the same time as
the claim.

Not committing the transaction before creating the claim resulted in
`Item.quantity_claimed` being set in an `INSERT` rather than in an
`UPDATE`. This led to incorrect treatment of the data.

Calling `db.session.commit()` after creating the test item allows
`Item.claim()` to behave in the same way as it would outside the tests.