Add tests for ProfileForm
These tests are all very similar to the `EventForm` tests added in
d81e27f.

There are also two trivial changes being made to the tests for events
forms. A line of whitespace is being added to make a test easier to read
and a comment is being added so that it's clear why a context is needed
for the tests.