Move context figure to conftest.py
None of the tests that need a request context actually use the context
directly. To save having to accept it as an argument, a test can be
marked to use a fixture with `pytest.mark.usefixtures()`.

Also, pytest is smart enough to load `conftest.py`. By defining the
`context` fixture here instead of in a fixtures module, it will
automatically be registered as a fixture that can be referenced in
`usefixtures()`. Otherwise `context` would need to be imported in order
to reference it.