Prevent slugify from adding trailing delimiters
If the first character, last character, or any consecutive characters of
a string being split is equal to the separator, the list will include
empty elements. Joining the list using a delimiter will result in the
delimiter appearing before the string, after the string, or repeated
throughout the string, respectively.

In the case of `slugify()`, this leaves to unintended results. Filtering
the empty results out of the list achieve the desired behavior.

Yay tests!