Use Factory Boy for test factories
Rather than reinventing the wheel for the factories used to generate
test objects, Factory Boy has the ability to do everything I was already
doing and then some. It even hooks into SQLAlchemy.

It's being added to both the tox and Travis CI setups as well as all the
test cases.