Add tests for the Unique validator
If it weren't for coverage I probably would have forgotten to add these
tests. There was a note in 68d46ab saying they'd be added at a later
time. At the time of that revision there was no way for the tests to
utilize the database. Now that they can, these tests are being added to
cover the functionality of the `Unique` validator.