Add change password page
Most of the work for this change is handled by Flask-Security, but a
custom template is being added to make it match the site better. Also,
rather than using Flask-Security's default behavior of sending a user to
the home page after they've changed their password, they will be
returned to their profile, which is where the 'Change password' link
lives.