Disable default messages on security templates
The default location for messages isn't very pretty on certain pages.
The templates here already had the code to display the messages in a
different location; they did not, however, have the code to prevent them
from displaying in the default location.

As a side note, the default location should probably be revisited, along
with the technique for displaying them.